NonNX/LINE計程車/requirements.txt
Flask==2.3.3
line-bot-sdk==3.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
APScheduler==3.10.4 